+++ vite/src/components/index.ts
@@ -4,5 +4,8 @@ export * from './wallet';
 // OFT token information components  
 export * from './oft';
 
+// OFT minting components
+export * from './mint';
+
 // Transfer/Send components
 export * from './send';
