+++ vite/src/components/mint/index.ts
@@ -1,0 +1,2 @@
+export { default as SolanaMintCard } from './SolanaMintCard';
+export { default as EvmMintCard } from './EvmMintCard';
